v0 code for better initial load page
