+++ flow/activities/flowable.go
@@ -545,6 +545,16 @@ func (a *FlowableActivity) GetQRepPartitions(ctx context.Context,
 	runUUID string,
 ) (*protos.QRepParitionResult, error) {
 	ctx = context.WithValue(ctx, shared.FlowNameKey, config.FlowJobName)
+	err := monitoring.InitializeQRepRun(
+		ctx,
+		a.CatalogPool,
+		config,
+		runUUID,
+		nil,
+	)
+	if err != nil {
+		return nil, err
+	}
 	srcConn, err := connectors.GetQRepPullConnector(ctx, config.SourcePeer)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get qrep pull connector: %w", err)
